+++ dataflowengine/src/test/scala/io/shiftleft/passes/dataflows/steps/DataFlowCodeToCpgFixture.scala
@@ -2,7 +2,8 @@ package io.shiftleft.passes.dataflows.steps
 
 import io.shiftleft.SerializedCpg
 import io.shiftleft.codepropertygraph.Cpg
-import io.shiftleft.layers.{DataFlowRunner, EnhancementRunner}
+import io.shiftleft.layers.EnhancementRunner
+import io.shiftleft.passes.dataflows.DataFlowRunner
 import io.shiftleft.semanticsloader.SemanticsLoader
 import io.shiftleft.testfixtures.{CodeToCpgFixture, LanguageFrontend}
